Revert "add pages layer"
