+++ src/App.scss
@@ -1,0 +1,9 @@
+.todo {
+    max-width: 1000px;
+    min-width: 600px;
+    margin: 20px auto 0px auto;
+    background: white;
+    padding: 24px;
+    border-radius: 4px;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.15);
+}
